[Feat] 게임상세조회 화면 - 게임삭제 버튼 추가 및 게임삭제 후 소속팀게임목록 조회를 위한  라우팅 파라미터 추가

+++ src/components/game/button/GameDeletionBtn.vue
@@ -18,15 +18,19 @@
 		data() {
 			return {
 				gameSeq: this.pGameSeq,
+				eventEmitName: 'delete-game',
 			};
 		},
 		methods: {
 			async deleteGame() {
+				if (!confirm('게임을 삭제하겠습니까?')) {
+					return;
+				}
 				const params = {
 					gameSeq: this.gameSeq,
 				};
 				await GameAPI.deleteGame(params);
-				this.$emit('delete-event-emit');
+				this.$emit(this.eventEmitName);
 			},
 		},
 	};
